SSMOBD2VirtualECU is minimally complete.  It essentially behaves like the TestMonitor, but emulates the communication behavior instead of just the parameters.  Consider this a checkpoint that could use a lot of cleanup.

+++ src/net/sourceforge/JDash/ecu/param/special/TimeParameter.java
@@ -79,7 +79,7 @@ public class TimeParameter extends Parameter
     public void marktime()
     {
         if(_init == 0) _init = System.currentTimeMillis();
-        marktime(System.currentTimeMillis() - (long)_init);
+        marktime(System.currentTimeMillis() - _init);
     }
 
     /******************************************************
